Move global examples to own package
