+++ src/test/java/fopbot/global/FillRectangle.java
@@ -1,4 +1,4 @@
-package fopbot.examples;
+package fopbot.global;
 
 
 import fopbot.Direction;
